Initial work for win64 support

+++ neo/idlib/Heap.cpp
@@ -1117,7 +1117,7 @@ void *Mem_Alloc16( const int size ) {
 	}
 	void *mem = mem_heap->Allocate16( size );
 	// make sure the memory is 16 byte aligned
-	assert( ( ((int)mem) & 15) == 0 );
+	assert( ( ((intptr_t)mem) & 15) == 0 );
 	return mem;
 }
 
@@ -1138,7 +1138,7 @@ void Mem_Free16( void *ptr ) {
 		return;
 	}
 	// make sure the memory is 16 byte aligned
-	assert( ( ((int)ptr) & 15) == 0 );
+	assert( ( ((intptr_t)ptr) & 15) == 0 );
  	mem_heap->Free16( ptr );
 }
